requirements.txt
google-cloud-storage==2.18.2
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
opencv-python==4.10.0.84
openai-whisper==20231117
moviepy==1.0.3
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
praw==7.7.1
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.4
tqdm==4.66.5
